pyproject.toml
[tool.poetry]
name = "gushare-backend"
version = "0.1.0"
description = "The backend application for the GUShare project."
authors = ["Christian Grossmüller", "Sören Grieger"]

[tool.isort]
profile = "black"
py_version = 38
multi_line_output = 3

[tool.black]
line-length = 80
target_version = ["py38"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "*"
django = "*"
gunicorn = "*"
django-environ = "*"
gevent = "*"
whitenoise = "*"
raven = "*"
djangorestframework = "*"
Pillow = "^9.4.0"
django-location-field = "^2.1.0"
black = "^23.1.0"
djoser = "*"
django-cors-headers = "^3.14.0"
dj-rest-auth = "1.0.9"
djangorestframework-simplejwt = "^5.2.2"
django-allauth = "^0.52.0"
django-filter = "^22.1"
drf-yasg = "^1.21.5"
unidecode = "^1.3.6"

[tool.poetry.dev-dependencies]
pytest = "*"
werkzeug = "*"
django-debug-toolbar = "*"
ipdb = "*"
pytest-django = "*"
django-extensions = "*"
pytest-isort = "*"
pytest-sugar = "*"
coverage = "*"
django-coverage-plugin = "*"
pytest-cov = "*"
codecov = "*"

[build-system]
# Should be the same as `$POETRY_VERSION`:
requires = ["poetry>=1.2.0"]
